// src/types.ts
export interface StudentUser {
  id: string;
  name: string;
  email: string;
  studentId: string;
  institute: string;
  department: string;
  semester: string;
  avatar: string;
  isVerified: boolean;
  joinedDate: string;
  rating: number;
  totalSales: number;
  totalPurchases: number;
  phone?: string;
  rollNumber?: string;
  isAdmin?: boolean;
  registrationNo?: string;
  verificationStatus?: VerificationStatus;
}
// ...
export type VerificationStatus = 'pending' | 'approved' | 'rejected';